+++ plugin-packs/postcss-preset-env/src/plugins/plugins-by-id.mjs
@@ -30,6 +30,7 @@ import postcssICUnit from '@csstools/postcss-ic-unit';
 import postcssImageSetFunction from 'postcss-image-set-function';
 import postcssIsPseudoClass from '@csstools/postcss-is-pseudo-class';
 import postcssLabFunction from 'postcss-lab-function';
+import postcssLightDarkFunction from '@csstools/postcss-light-dark-function';
 import postcssLogicalOverflow from '@csstools/postcss-logical-overflow';
 import postcssLogicalOverscrollBehavor from '@csstools/postcss-logical-overscroll-behavior';
 import postcssLogical from 'postcss-logical';
@@ -91,6 +92,7 @@ export const pluginsById = new Map(
 		['image-set-function', postcssImageSetFunction],
 		['is-pseudo-class', postcssIsPseudoClass],
 		['lab-function', postcssLabFunction],
+		['light-dark-function', postcssLightDarkFunction],
 		['logical-overflow', postcssLogicalOverflow],
 		['logical-overscroll-behavior', postcssLogicalOverscrollBehavor],
 		['logical-properties-and-values', postcssLogical],